ETL/requirements.txt
asn1crypto==1.5.1
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
cryptography==42.0.5
filelock==3.13.1
idna==3.7
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
numexpr==2.8.7
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pip==24.0
platformdirs==3.10.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
setuptools==69.5.1
six==1.16.0
snowflake-connector-python==3.10.1
sortedcontainers==2.4.0
tomlkit==0.11.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==1.26.18
wheel==0.43.0
win-inet-pton==1.1.0
